+++ gwt-uibinder-example/gwt-uibinder-example-client/src/main/module.gwt.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright 2018 Vertispan LLC
+
+    Licensed under the Apache License, Version 2.0 (the "License"); you may not
+    use this file except in compliance with the License. You may obtain a copy of
+    the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+    License for the specific language governing permissions and limitations under
+    the License.
+
+-->
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit 2.8.2//EN"
+  "http://gwtproject.org/doctype/2.8.2/gwt-module.dtd">
+<module>
+  <inherits name="com.google.gwt.user.User" />
+  <inherits name="com.google.gwt.user.theme.clean.Clean" />
+
+  <!-- to be replaced -->
+  <inherits name="com.google.gwt.uibinder.UiBinder"/>
+
+  <!-- new UiBinder -->
+  <inherits name="org.gwtproject.uibinder.UiBinder"/>
+
+  <inherits name="com.google.common.util.concurrent.Concurrent"/>
+  <inherits name="dagger.Dagger"/>
+  
+  <source path="" />
+
+  <entry-point class="org.gwtproject.uibinder.example.Application" />
+
+  <!-- Only support recent browsers -->
+  <set-property name="user.agent" value="ie10,gecko1_8,safari" />
+</module>
