+++ gwt-uibinder-example/gwt-uibinder-example-client/src/main/resources/org/gwtproject/uibinder/example/view/SimpleFormView.ui.xml
