+++ gwt-uibinder-example/gwt-uibinder-example-server/src/main/jettyconf/context.xml
@@ -1,0 +1,6 @@
+<Configure class="org.eclipse.jetty.webapp.WebAppContext">
+  <Call name="setInitParameter">
+    <Arg>org.eclipse.jetty.servlet.Default.useFileMappedBuffer</Arg>
+    <Arg>false</Arg>
+  </Call>
+</Configure>
