+++ gwt-uibinder-example/gwt-uibinder-example-server/src/main/tomcatconf/context.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Context reloadable="true">
+
+  <Resources className="org.apache.naming.resources.VirtualDirContext"
+    extraResourcePaths="/=${basedir}/../target/gwt/launcherDir/" />
+</Context>
