--- gwt-uibinder-test/pom.xml
@@ -1,130 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.gwtproject</groupId>
-    <artifactId>gwt-uibinder</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
-  </parent>
-
-
-  <artifactId>gwt-uibinder-test</artifactId>
-  <name>GWT UiBinder Testbed</name>
-  <description>Intended to be a test module for GWT UiBinder</description>
-
-  <packaging>gwt-app</packaging>
-
-  <properties>
-    <dagger.version>2.15</dagger.version>
-    <license.orgName>Vertispan LLC</license.orgName>
-  </properties>
-
-  <build>
-    <resources>
-      <resource>
-        <directory>src/main/java</directory>
-      </resource>
-      <resource>
-        <directory>src/main/resources</directory>
-      </resource>
-      <resource>
-        <directory>target/generated-sources/annotations</directory>
-      </resource>
-    </resources>
-
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <groupId>com.mycila</groupId>
-        <artifactId>license-maven-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <groupId>net.ltgt.gwt.maven</groupId>
-        <artifactId>gwt-maven-plugin</artifactId>
-        <configuration>
-          <moduleName>org.gwtproject.uibinder.test.GwtUiBinderTest</moduleName>
-          <moduleShortName>uibindertest</moduleShortName>
-          <skipModule>true</skipModule>
-        </configuration>
-      </plugin>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-war-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <version>${project.version}</version>
-      <artifactId>gwt-uibinder-client</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>${project.groupId}</groupId>
-      <version>${project.version}</version>
-      <artifactId>gwt-uibinder-processor</artifactId>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.gwt</groupId>
-      <artifactId>gwt-user</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.gwt</groupId>
-      <artifactId>gwt-dev</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.dagger</groupId>
-      <artifactId>dagger-gwt</artifactId>
-      <version>${dagger.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.dagger</groupId>
-      <artifactId>dagger-producers</artifactId>
-      <version>${dagger.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.dagger</groupId>
-      <artifactId>dagger-producers</artifactId>
-      <version>${dagger.version}</version>
-      <classifier>sources</classifier>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.dagger</groupId>
-      <artifactId>dagger-compiler</artifactId>
-      <version>${dagger.version}</version>
-      <scope>provided</scope>
-      <optional>true</optional>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <version>${guava.version}</version>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava-gwt</artifactId>
-      <version>${guava.version}</version>
-    </dependency>
-
-  </dependencies>
-
-</project>
