--- gwt-uibinder-test/src/main/webapp/index.html
@@ -1,20 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset=UTF-8>
-  <title>GWT UiBinder Test</title>
-
-  <script src="uibindertest/uibindertest.nocache.js"></script>
-</head>
-<body>
-
-<noscript>
-  <div
-      style="width: 22em; position: absolute; left: 50%; margin-left: -11em; color: red; background-color: white; border: 1px solid red; padding: 4px; font-family: sans-serif">
-    Your web browser must have JavaScript enabled
-    in order for this application to display correctly.
-  </div>
-</noscript>
-
-</body>
-</html>
